+++ projects/fhi-angular-components/src/public-api.ts
@@ -7,7 +7,6 @@ export * from './lib/fhi-angular-components.module';
 export * from './lib/fhi-autosuggest/fhi-autosuggest.component';
 
 export * from './lib/fhi-multiselect/fhi-multiselect.component';
-export * from './lib/fhi-multiselect/fhi-multiselect.module';
 
 export * from './lib/fhi-table/fhi-table.component';
 export * from './lib/fhi-table/row-components/fhi-table-row-expander/fhi-table-row-expander.component';